places-setup: check if exists before deleting

+++ lib/Command/PlacesSetup.php
@@ -395,6 +395,10 @@ class PlacesSetup extends Command
 
     protected function ensureDeleted(string $filename)
     {
+        if (!file_exists($filename)) {
+            return;
+        }
+
         unlink($filename);
         if (file_exists($filename)) {
             $this->output->writeln('<error>Failed to delete data file</error>');
